requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
psycopg2-binary==2.9.1
pydantic==1.8.2
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
